requirements.txt
chromadb
openai
langchain
langchain-openai
langchain-community
langchain-chroma
python-dotenv